% Generated by roxygen2 (4.0.1): do not edit by hand
\name{betaCoef}
\alias{betaCoef}
\title{Regression coefficients from full model fit}
\usage{
betaCoef(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{betaCoef} returns the regression coefficients.
}
\description{
Access the full model fitted coefficients of an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{betaCoef}

