% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{betaCoef}
\alias{betaCoef}
\alias{betaCoef,edgeFit-method}
\title{Regression coefficients from full model fit}
\usage{
betaCoef(object)

\S4method{betaCoef}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{betaCoef} returns the regression coefficients.
}
\description{
Access the full model fitted coefficients of an
\code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{betaCoef}

