% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/form_models.R
\name{build_models}
\alias{build_models}
\title{Generate alternative and null hypothesis in deSet object}
\usage{
build_models(data, cov, full.model = NULL, null.model = NULL, ind = NULL)
}
\arguments{
\item{data}{\code{matrix}: gene expression data.}

\item{cov}{\code{data.frame}: biological covariates.}

\item{full.model}{\code{formula}: the adjusment and biological variables of
interest.}

\item{null.model}{\code{formula}: the adjustment variables.}

\item{ind}{\code{factor}: individuals sampled in the study.}
}
\value{
\code{build_models} returns an \code{\linkS4class{deSet}} object
with the following slots assigned:
\describe{
   \item{\code{full.model:}}{full model equation}
   \item{\code{null.model:}}{null model equation}
   \item{\code{full.matrix:}}{full model in matrix form}
   \item{\code{null.matirx:}}{null model in matrix form}
   \item{\code{individual:}}{individuals in experiment (factor)}
   \item{\code{ExpressionSet:}}{inherits ExpressionSet object (assayData,
   phenoData) created in function}
 }
}
\description{
\code{build_models} is a function to create a deSet without an
ExpressionSet. Alternative and null models are inputting by users.
}
\examples{
# create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
age <- kidney$age
sex <- kidney$sex
kidexpr <- kidney$kidexpr
cov <- data.frame(sex = sex, age = age)

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df=4)

# create deSet object from data
de_obj <- build_models(data = kidexpr, cov = cov, null.model = null.model,
full.model = full.model)
}
\author{
John Storey, Andy Bass
}
\seealso{
\code{\link{deSet}}
}

