% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{deSet-class}
\alias{deSet-class}
\title{deSet class}
\description{
The deSet class was designed in order to complement the
\code{\link{ExpressionSet}} class. While the \code{ExpressionSet} class
contains information about the experiment, the deSet class
contains both experimental information and additional information relevant
to differential expression analysis.
}
\details{
The deSet object is required for the following functions:
}
\section{Slots}{

\describe{
\item{\code{null.model}}{\code{formula}: contains the adjustment variables in the
experiment.}

\item{\code{full.model}}{\code{formula}: contains the adjustment variables and the
biological variables of interest.}

\item{\code{null.matrix}}{\code{matrix}: the null model as a matrix.}

\item{\code{full.matrix}}{\code{matrix}: the full model as a matrix.}

\item{\code{individual}}{\code{factor}: containing information on individuals
sampled in the experiment.}

\item{\code{qvalueObj}}{S3 class \code{qvalue}: containing qvalue object.
 See \code{\link{qvalue}} for additional details.}
}}
\note{
The format for the model inputs are
\itemize{
 \item full.model: adjustment variables + biological variables
 \item null.model: adjustment variables
}
The deSet object is created by either using \code{\link{deSet}},
\code{\link{build_models}}, or \code{\link{build_study}} functions.
The qvalueObj is the slot of interest and is determined by either
using the \code{odp} or the \code{lrt} function.
}
\section{Methods}{

 \describe{
 \item{\code{as(ExpressionSet, "deSet")}}{Coerce objects of ExpressionSet
 to deSet}
 \item{\code{lrt(deSet, ...)}}{Likelihood ratio test}
 \item{\code{odp(deSet, ...)}}{Optimal discovery procedure}
 \item{\code{kl_clust(deSet, ...)}}{Clustering parameters for modular
 optimal discovery procedure (mODP) method}
 \item{\code{fit_models(deSet, ...)}}{Linear regression for genes based on
             null and full models}
 \item{\code{apply_qvalue(deSet, ...)}}{Implements qvalue function on
 deSet object}
 \item{\code{apply_snm(deSet, ...)}}{Implement surpervised normalization of
  microarrays on gene expression matrix on deSet object.}
 \item{\code{apply_sva(deSet, ...)}}{Estimate surrogate variables and adds
 them to null/full models in a deSet object}
 \item{\code{fullMatrix(deSet)}}{Access and set full matrix from
 deSet object}
 \item{\code{nullMatrix(deSet)}}{Access and set null matrix from
 deSet object}
 \item{\code{fullModel(deSet)}}{Access and set full model from
 deSet object}
 \item{\code{nullModel(deSet)}}{Access and set null model from
 deSet object}
 \item{\code{individual(deSet)}}{Set individual slot from deSet object}
 \item{\code{qvalueObj(deSet)}}{Access qvalue object from deSet object.
 See \code{\link{qvalue}}.}
 \item{\code{validObject(deSet)}}{Check validity of deSet object.}
 }
}
\author{
John Storey, Jeffrey Leek, Andrew Bass
}
\seealso{
\code{\link{deSet}}
}

