% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{edgeFit-class}
\alias{edgeFit-class}
\title{edgeFit class}
\description{
Object returned from \code{edgeFit-methods} function containing information
on the model fits for the experiment.
}
\details{
Object contains output from fitting a linear model to each gene for both the
full and null models. The stat.type caneither be "odp" or "lrt" depending on
which statistic is of interest to generate the p-values.
}
\note{
The edgeFit object can be the inputs for the \code{lrt}, \code{odp}
and \code{klClust} function but the object is automatically generated if not
specified.
}
\section{Slots}{

 \describe{
   \item{\code{fit.full}:}{Matrix of class \code{"matrix"},
                           containing fitted values for full model.}
   \item{\code{fit.null}:}{Matrix of class \code{"matrix"},
                           containing fitted values for null model.}
   \item{\code{res.full}:}{Matrix of class \code{"matrix"},
                           containing residuals for full model.}
   \item{\code{res.null}:}{Matrix of class \code{"matrix"},
                           containing residuals for null model.}
   \item{\code{dH.full}:}{Vector of class \code{"vector"},
                          containing diagonal elements in projection matrix
                          for the full model.}
   \item{\code{beta.coef}:}{Matrix of class \code{"matrix"},
                          containing linear fitted coefficients for full model.}
   \item{\code{stat.type}:}{String of class \code{"string"},
                            containing information on the statistic of
                            interest. Currently, the only options are lrt
                            and odp.}
 }
}

\section{Methods}{

 \describe{
 \item{\code{modelFits(edgeFit, ...)}}{Fitted models of edgeFit object}
 \item{\code{fitNull(edgeFit)}}{Fitted data from null model}
 \item{\code{fitFull(edgeFit)}}{Fitted data from full model}
 \item{\code{resNull(edgeFit)}}{Null residuals from null model}
 \item{\code{resFull(edgeFit)}}{Full residuals from full model}
 \item{\code{betaCoef(edgeFit)}}{Beta coefficients in linear model}
 \item{\code{sType(edgeFit)}}{Statistic type for model fitting utilized in function}
 }
}
\author{
John Storey, Jeffrey T. Leek, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{edgeFit-class}

