% Generated by roxygen2 (4.0.1): do not edit by hand
\name{edgeFit}
\alias{edgeFit}
\title{Linear Regression for full and null models}
\usage{
edgeFit(object, stat.type = c("lrt", "odp"))
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object.}

\item{stat.type}{character- type of statistic to be used. Either "lrt" or "odp".
Default is "lrt".}
}
\value{
\code{edgeFit} returns an \code{\linkS4class{edgeFit}} object.
}
\description{
\code{edgeFit} fits a linear model to each gene. Model fits can be either statistic type
odp (Optimal Discovery Procedure) or lrt (Likelihood Ratio Test).
}
\details{
If individual factors exists from \code{edgeSet} object, they are removed from the model to remove
the effects from individuals. If ODP method is implemented then the null model
is removed from the full model (see Storey 2007). Statistics of interest for functions \code{lrt}, \code{odp} and
\code{klClust} are the elements of the projection matrix, the fitted
expression data, the regression coefficients and the residuals of the expression data for both the null
and full models.
}
\note{
\code{edgeFit} does not have to be called by the user to use
\code{odp}, \code{klClust} or \code{lrt} as it is an optional input and is
implemented in the methods if not specified. edgeFit object can be created
by the user if a different statistical implementation is required.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex[kidney$tissue == "c"]
age <- kidney$age[kidney$tissue == "c"]
kidexpr <- log(kidney$kidexpr[, kidney$tissue == "c"] + 10)
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex + ns(age, df=3, intercept=FALSE)

# edgeSet object
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# Retrieve statistics from linear regression for each gene
ef.lrt <- edgeFit(edgeObj, stat.type="lrt") # lrt method
ef.odp <- edgeFit(edgeObj, stat.type="odp") # odp method
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to simultaneous significance testing. Journal of the Royal Statistical Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for large-scale significance testing, with applications to comparative microarray experiments. Biostatistics, 8: 414-432.

Storey JD, Xiao W, Leek JT, Tompkins RG, and Davis RW. (2005) Significance analysis of time course microarray experiments. Proceedings of the National Academy of Sciences, 102: 12837-12842.
}
\seealso{
\code{\linkS4class{edgeFit}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeFit}

