% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{edgeFit}
\alias{edgeFit}
\alias{edgeFit,edgeSet-method}
\title{Least squares model fit for full and null models}
\usage{
edgeFit(object, stat.type = c("lrt", "odp"))

\S4method{edgeFit}{edgeSet}(object, stat.type = c("lrt", "odp"))
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object.}

\item{stat.type}{character- type of statistic to be used. Either "lrt" or
"odp". Default is "lrt".}
}
\value{
\code{edgeFit} returns an \code{\linkS4class{edgeFit}} object.
}
\description{
\code{edgeFit} fits a linear model to each gene by least squares. Model fits can be either
statistic type odp (optimal discovery procedure) or lrt
(likelihood ratio test).
}
\details{
If individual factors exists from \code{edgeSet} object, they are
adjusted from the data to remove the effects from repeated individuals. If
"odp" method is implemented then the null model is removed from the full
model (see Storey 2007). The statistic of interest for \code{lrt},
\code{odp} and \code{klClust} are the elements of the projection matrix,
the fitted expression data, the regression coefficients and the residuals
of the expression data for both the null and full models.
}
\note{
\code{edgeFit} does not have to be called by the user to use
\code{odp}, \code{klClust} or \code{lrt} as it is an optional input and is
implemented in the methods. edgeFit object can be created
by the user if a different statistical implementation is required.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# retrieve statistics from linear regression for each gene
ef_lrt <- edgeFit(edge_obj, stat.type = "lrt") # lrt method
ef_odp <- edgeFit(edge_obj, stat.type = "odp") # odp method
}
\author{
John Storey
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to
simultaneous significance testing. Journal of the Royal Statistical
Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for
large-scale significance testing, with applications to comparative
microarray experiments. Biostatistics, 8: 414-432.

Storey JD, Xiao W, Leek JT, Tompkins RG, and Davis RW. (2005) Significance
analysis of time course microarray experiments. Proceedings of the National
Academy of Sciences, 102: 12837-12842.
}
\seealso{
\code{\linkS4class{edgeFit}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeFit}

