% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edgeModel.R
\name{edgeModel}
\alias{edgeModel}
\title{Generate alternative and null hypothesis in edgeSet object}
\usage{
edgeModel(data, cov, full.model = NULL, null.model = NULL, ind = NULL)
}
\arguments{
\item{data}{matrix- Gene expression data.}

\item{cov}{data.frame- Biological covariates.}

\item{full.model}{formula- Full model of the experiment.}

\item{null.model}{formula- Null model of the experiment.}

\item{ind}{factor. Individuals sampled in the study.}
}
\value{
\code{edgeModel} returns an \code{\linkS4class{edgeSet}} object
with the following slots assigned:
\describe{
   \item{\code{full.model:}}{alternative model equation}
   \item{\code{null.model:}}{null model equation}
   \item{\code{full.matrix:}}{alternative model in matrix form}
   \item{\code{null.matirx:}}{null model in matrix form}
   \item{\code{individual:}}{individuals in experiment (factor)}
   \item{\code{ExpressionSet:}}{inherits ExpressionSet object (assayData,
   phenoData) created in function}
 }
}
\description{
\code{edgeModel} is a function to create an edge object without an
ExpressionSet. Alternative and null models are inputting by users.
}
\examples{
# create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df=4)

# create edgeSet object from data
edgeObj <- edgeModel(data = kidexpr, cov = cov, null.model = null.model,
full.model = full.model)
}
\author{
John Storey, Andy Bass
}
\seealso{
\code{\link{edgeSet}}
}

