% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{edgeQvalue}
\alias{edgeQvalue}
\title{Estimate the q-values for a given set of p-values in edgeSet object}
\usage{
edgeQvalue(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{...}{Additional arguments for \code{\link{qvalue}}}
}
\value{
\code{edgeQvalue} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{qvalue} on edgeSet object based on p-values determined from functions
\code{lrt} or \code{odp}
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# Run lrt (or odp) and edgeQvalue
edge.lrt <- lrt(edgeObj)
edge.lrt.new <- edgeQvalue(edge.lrt, fdr.level=0.05, pi0.method="bootstrap", adj=1.2)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD and Tibshirani R. (2003) Statistical significance for genome-wide studies. Proceedings of the National Academy of Sciences, 100: 9440-9445
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and \code{\link{lrt}}
}
\keyword{edgeQvalue}

