% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{edgeQvalue}
\alias{edgeQvalue}
\alias{edgeQvalue,edgeSet-method}
\title{Estimate the q-values for a given set of p-values in an edgeSet object}
\usage{
edgeQvalue(object, ...)

\S4method{edgeQvalue}{edgeSet}(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{...}{Additional arguments for \code{\link{qvalue}}}
}
\value{
\code{edgeQvalue} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{qvalue} on edgeSet object based on p-values determined from
\code{lrt} or \code{odp}
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# Run lrt (or odp) and edgeQvalue
edge_lrt <- lrt(edge_obj)
edge_lrt <- edgeQvalue(edge_lrt, fdr.level = 0.05,
pi0.method = "bootstrap", adj=1.2)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD and Tibshirani R. (2003) Statistical significance for
genome-wide studies. Proceedings of the National Academy of Sciences, 100: 9440-9445
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeQvalue}

