% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{edgeSet}
\alias{edgeSet}
\title{Creates object of class edgeSet}
\usage{
edgeSet(object, full.model, null.model, individual = NULL)
}
\arguments{
\item{object}{\code{\link{ExpressionSet}} object}

\item{full.model}{formula- full model for experiment data.}

\item{null.model}{formula- null model for experiment data.}

\item{individual}{factor- information on individuals in experiment.}
}
\value{
\code{edgeSet} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Creates an edgeSet object that extends ExpressionSet.
}
\note{
It is essential that the null and full models have the same variables
as the ExpressionSet phenoType column names.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# Optionally add individuals to experiment for kidney data not applicable
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel, ind=factor(1:72))
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and \code{\link{lrt}}
}
\keyword{edgeSet}

