% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ExpressionSet-methods.R
\docType{methods}
\name{edgeSet}
\alias{edgeSet}
\alias{edgeSet,ExpressionSet-method}
\title{Creates object of class edgeSet}
\usage{
edgeSet(object, full.model, null.model, individual = NULL)

\S4method{edgeSet}{ExpressionSet}(object, full.model, null.model,
  individual = NULL)
}
\arguments{
\item{object}{\code{\link{ExpressionSet}} object}

\item{full.model}{formula- full model for experiment data.}

\item{null.model}{formula- null model for experiment data.}

\item{individual}{factor- information on individuals in experiment.}
}
\value{
\code{edgeSet} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Creates an edgeSet object that extends ExpressionSet.
}
\note{
It is essential that the null and full models have the same variables
as the ExpressionSet phenoType column names.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# Optionally add individuals to experiment for kidney data not applicable
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model, ind = factor(1:72))
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeSet}

