% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edgeSNM}
\alias{edgeSNM}
\title{Supervised normalization of data in edgeSet object}
\usage{
edgeSNM(object, int.var, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{int.var}{data frame- intensity-dependent effects.}

\item{...}{Additional arguments for \code{\link{snm}}}
}
\value{
\code{edgeSNM} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{snm} on edgeSet object based on the null and full models in \code{\linkS4class{edgeSet}}
}
\examples{
# Simulate data
require(snm)
singleChannel <- sim.singleChannel(12345)
# Create edgeSet object using edgeModel (can use ExpressionSet see manual)
grp = singleChannel$bio.var[,2]
adj.var = data.frame(singleChannel$adj.var[,-1])
Create edgeSet object using edgeModel (can use ExpressionSet -- see manual)
edgeObj <- edgeModel(data=singleChannel$raw.data, sampling="static", grp=grp, adj.var=adj.var)

# Run SNM using intensity-dependent adjustment variable
nEdgeObj <- edgeSNM(edgeObj, int.var=singleChannel$int.var, verbose=FALSE, num.iter=1)
}
\author{
John Storey, Andrew Bass
}
\references{
Mechan BH, Nelson PS, Storey JD. Supervised normalization of microarrays. Bioinformatics 2010;26:1308-15
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and \code{\link{lrt}}
}
\keyword{edgeSNM}

