% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{edgeSNM}
\alias{edgeSNM}
\alias{edgeSNM,edgeSet-method}
\title{Supervised normalization of data in edge}
\usage{
edgeSNM(object, int.var, ...)

\S4method{edgeSNM}{edgeSet}(object, int.var, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{int.var}{data frame- intensity-dependent effects.}

\item{...}{Additional arguments for \code{\link{snm}}}
}
\value{
\code{edgeSNM} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{snm} on edgeSet object based on the null and full models in
\code{\linkS4class{edgeSet}}. See \code{\link{snm}} for additional details
on the algorithm.
}
\examples{
# simulate data
require(snm)
singleChannel <- sim.singleChannel(12345)
data <- singleChannel$raw.data

# create edgeSet object using edgeModel (can use ExpressionSet see manual)
cov <- data.frame(grp = singleChannel$bio.var[,2])
full.model <- ~grp
null.model <- ~1
# Create edgeSet object using edgeModel
edge_obj <- edgeModel(data = data, altMod = full.model,
nullMod = null.model, cov = cov)

# Run SNM using intensity-dependent adjustment variable
edge_snm <- edgeSNM(edge_obj, int.var = singleChannel$int.var,
verbose = FALSE, num.iter = 1)
}
\author{
John Storey, Andrew Bass
}
\references{
Mechan BH, Nelson PS, Storey JD. Supervised normalization of microarrays.
Bioinformatics 2010;26:1308-15
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeSNM}

