% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{edgeSVA}
\alias{edgeSVA}
\alias{edgeSVA,edgeSet-method}
\title{Estimate surrogate variables}
\usage{
edgeSVA(object, ...)

\S4method{edgeSVA}{edgeSet}(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{...}{Additional arguments for \code{\link{sva}}}
}
\value{
\code{edgeSVA} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{sva} on edgeSet object based on the null and full models in
\code{\linkS4class{edgeSet}}. See \code{\link{sva}} for additional details.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# run surrogate variable analysis
edge_sva <- edgeSVA(edge_obj)

# run odp/lrt with surrogate variables added
edge_odp <- odp(edge_sva, bs.its = 20)
summary(edge_odp)
}
\author{
John Storey, Andrew Bass
}
\references{
Leek JT, Storey JD (2007) Capturing Heterogeneity in Gene Expression
Studies by Surrogate Variable Analysis. PLoS Genet 3(9): e161.
doi:10.1371/journal.pgen.0030161
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and
\code{\link{lrt}}
}
\keyword{edgeSVA}

