% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fitFull}
\alias{fitFull}
\title{Fitted data from the full model}
\usage{
fitFull(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{fitFull} returns a matrix of fitted values from full model.
}
\description{
Access the fitted data from the full model in an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{fitFull}

