% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{fitFull}
\alias{fitFull}
\alias{fitFull,edgeFit-method}
\title{Fitted data from the full model}
\usage{
fitFull(object)

\S4method{fitFull}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{fitFull} returns a matrix of fitted values from full model.
}
\description{
Access the fitted data from the full model in an
\code{\linkS4class{edgeFit}} object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# run edgeFit to get model fits
edge_fit <- edgeFit(edge_obj)

# extract fitted values for full model
fitted_full <- fitFull(edge_fit)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{fitFull}

