% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitNull}
\alias{fitNull}
\title{Fitted data from the null model}
\usage{
fitNull(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{fitFull} returns a matrix of fitted values from null model.
}
\description{
Access the fitted data from the null model in an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{fitNull}

