% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fullMatrix}
\alias{fullMatrix}
\alias{fullMatrix<-}
\title{Matrix representation of full model}
\usage{
fullMatrix(object)

fullMatrix(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{matrix}: full model matrix where the columns are the covariates amd rows are observations}
}
\value{
\code{fullMatrix} returns the value of the full model matrix.
}
\description{
These generic functions access and set the full matrix for
\code{\linkS4class{edgeSet}} object.
}
\author{
Andrew Bass, John Storey
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{fullModel}}
}
\keyword{fullMatrix,}
\keyword{fullMatrix<-}

