% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fullModel}
\alias{fullModel}
\alias{fullModel<-}
\title{Full model equation from edgeSet object}
\usage{
fullModel(object)

fullModel(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{formula}: The experiment design for the full model.}
}
\value{
\code{fullModel} returns the formula for the full model.
}
\description{
These generic functions access and set the full model for
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{fullModel,}
\keyword{fullModel<-}

