% Generated by roxygen2 (4.0.1): do not edit by hand
\name{individual}
\alias{individual}
\alias{individual<-}
\title{Individuals utilized in experiment}
\usage{
individual(object)

individual(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{factor}: individual identifier for each observation}
}
\value{
\code{individual} returns information regarding individuals
in the experiment.
}
\description{
These generic functions access and set the individual slot in
\code{\linkS4class{edgeSet}}.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{individual,}
\keyword{individual<-}

