% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{individual}
\alias{individual}
\alias{individual,edgeSet-method}
\alias{individual<-}
\alias{individual<-,edgeSet-method}
\title{Individuals sampled in experiment}
\usage{
individual(object)

individual(object) <- value

\S4method{individual}{edgeSet}(object) <- value

\S4method{individual}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{factor}: individual identifier for each observation}
}
\value{
\code{individual} returns information regarding individuals
in the experiment.
}
\description{
These generic functions access and set the individual slot in
\code{\linkS4class{edgeSet}}.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{individual,}
\keyword{individual<-}

