% Generated by roxygen2 (4.0.1): do not edit by hand
\name{klClust}
\alias{klClust}
\title{Clustering (mODP) method on edgeSet object}
\usage{
klClust(object, obj.edgeFit = NULL, n.mods = 50, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object.}

\item{n.mods}{numeric- number of clusters.}

\item{\dots}{additional parameters}
}
\value{
\code{klClust} returns a list:
\itemize{
\item {mu.full: mean of clusters from full model}
\item {mu.null: mean of clusters from null model}
\item {sig.full: sd of clusters from full model}
\item {sig.null: sd of clusters from null model}
\item {n.per.mod: total members in each cluster}
\item {clustMembers: members for each gene}

}
}
\description{
\code{klClust} is an implementation of mODP that assigns genes to modules based
off of the Kullback-Leibler distance.
}
\details{
mODP utilizes a k-means clustering algorithm where genes are assigned
to a cluster based on the Kullback-Leiber distance. Each gene is assigned an
module-average parameter to calculate the ODP score (See Woo, Leek and Storey 2010 for more details).
The mODP and full ODP produce near exact results but mODP has the advantage of being computationally
feasible.
}
\note{
The results are generally insensitive to the number of modules after a
certain threshold of about K>=50.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# ODP method
edge.clust <- klClust(edgeObj)

# Change the number of clusters
edge.clust <- klClust(edgeObj, n.mods=10)

# Input an edgeFit object or ODP parameters... not necessary
edge.fit <- edgeFit(edgeObj, stat.type="odp")
edge.clust.ef <- klClust(edgeObj, obj.edgeFit=edge.fit)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to simultaneous significance testing. Journal of the Royal Statistical Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for large-scale significance testing, with applications to comparative microarray experiments. Biostatistics, 8: 414-432.

Woo S, Leek JT, Storey JD (2010) A computationally efficient modular optimal discovery procedure. Bioinformatics, 27(4): 509-515.
}
\seealso{
\code{\link{odp}}, \code{\link{edgeSet}} and \code{\link{edgeFit}}
}
\keyword{klClust}

