% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{klClust}
\alias{klClust}
\alias{klClust,edgeSet,edgeFit-method}
\alias{klClust,edgeSet,missing-method}
\title{Clustering (mODP) method}
\usage{
klClust(object, obj.edgeFit = NULL, n.mods = 50, ...)

\S4method{klClust}{edgeSet,missing}(object, obj.edgeFit = NULL, n.mods = 50,
  ...)

\S4method{klClust}{edgeSet,edgeFit}(object, obj.edgeFit = NULL, n.mods = 50,
  ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object.}

\item{n.mods}{numeric- number of clusters.}

\item{\dots}{additional parameters}
}
\value{
\code{klClust} returns a list:
\itemize{
  \item {mu.full: mean of clusters from full model}
  \item {mu.null: mean of clusters from null model}
  \item {sig.full: sd of clusters from full model}
  \item {sig.null: sd of clusters from null model}
  \item {n.per.mod: total members in each cluster}
  \item {clustMembers: members for each gene}
}
}
\description{
\code{klClust} is an implementation of mODP that assigns genes to modules
based off of the Kullback-Leibler distance.
}
\details{
mODP utilizes a k-means clustering algorithm where genes are
assigned to a cluster based on the Kullback-Leiber distance. Each gene is
assigned an module-average parameter to calculate the ODP score (See Woo,
Leek and Storey 2010 for more details). The mODP and full ODP produce near
exact results but mODP has the advantage of being computationally
feasible.
}
\note{
The results are generally insensitive to the number of modules after a
certain threshold of about n.mods>=50.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# ODP method
edge_clust <- klClust(edge_obj)

# Change the number of clusters
edge_clust <- klClust(edge_obj, n.mods = 10)

# Input an edgeFit object or ODP parameters... not necessary
edge_fit <- edgeFit(edge_obj, stat.type = "odp")
edge_clust <- klClust(edge_obj, obj.edgeFit = edge_fit)
}
\author{
John Storey, Jeffrey Leek
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to
simultaneous significance testing. Journal of the Royal Statistical
Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for
large-scale significance testing, with applications to comparative
microarray experiments. Biostatistics, 8: 414-432.

Woo S, Leek JT, Storey JD (2010) A computationally efficient modular optimal
 discovery procedure. Bioinformatics, 27(4): 509-515.
}
\seealso{
\code{\link{odp}}, \code{\link{edgeSet}} and \code{\link{edgeFit}}
}
\keyword{klClust}

