% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lrt}
\alias{lrt}
\title{Get model equations from edgeFit object}
\usage{
lrt(object, obj.edgeFit, nullDistn = c("normal", "bootstrap"), bs.its = 100,
  seed = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object- S4 class}

\item{nullDistn}{character- either "normal" or "bootstrap", If "normal",
use p-values from ANOVA else if "bootstrap" p-values will
be determined from function \code{\link{empPvals}}.
Default is "normal".}

\item{bs.its}{numeric- number of null statistics generated (only applicable
for "bootstrap" method. Default is 100.}

\item{seed}{numeric- set the seed value.}

\item{verbose}{boolean- print iterations for bootstrap method. Default is TRUE.}

\item{...}{Additional arguments for \code{qvalue} and \code{empPvals} function.}
}
\value{
\code{lrt} returns a \code{\linkS4class{edgeSet}} object with slot qvalueObj as \code{\link{qvalue}} object.
}
\description{
Performs F-test (Generalized Likelihood Ratio)
}
\details{
\code{lrt} performs Generalized Likelihood Ratio on the full and null models to determine p-values (either from a F-distribution or
from simulated null test statistics). If the null distribution is calculated using "bootstrap" then
residuals from the alternative model are resampled and added to the null model to simulate
a distribution where there is no differential expression. Otherwise, the default
input is "normal" and the assumption is that the dataset follows an F-distribution. Note that
if "normal" is chosen then the sum of squares should be statistically independent
and genes should be normally distributed with the same variance. The \code{qvalue} function
is used to retrieve the \code{qvalue} object. See \code{\link{qvalue}} for more information.
}
\note{
Fits a linear regression to each gene from function
\code{\link{edgeFit}} and then performs an ANOVA F-test on the null and full
model fits. If nullDistn "normal" is chosen then the F-distribution is used to determine
p-values else if "bootstrap" is chosen then empirical p-values will be
determined from \code{\link{qvalue}} package (see \code{\link{empPvals}}).
To retrieve linear regression results, see \code{\link{edgeFit}}.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex[kidney$tissue == "c"]
age <- kidney$age[kidney$tissue == "c"]
kidexpr <- log(kidney$kidexpr[, kidney$tissue == "c"] + 10)
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# lrt method
edge.lrt <- lrt(edgeObj, nullDistn="normal")

# To generate pvalues from bootstrap
edge.lrt.bs <- lrt(edgeObj, nullDistn="bootstrap")

# Input an edgeFit object but not necessary
edge.fit <- edgeFit(edgeObj, stat.type="lrt")
edge.lrt.ef <- lrt(edgeObj, obj.edgeFit=edge.fit)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD, Xiao W, Leek JT, Tompkins RG, and Davis RW. (2005) Significance analysis of time course microarray experiments. Proceedings of the National Academy of Sciences, 102: 12837-12842.
}
\seealso{
\code{\link{edgeSet}}, \code{\link{odp}}, \code{\link{edgeFit}}
}
\keyword{lrt}

