% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modelFits}
\alias{modelFits}
\title{Fitted models for an \code{\linkS4class{edgeFit}} object.}
\usage{
modelFits(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}} object}
}
\value{
\code{modelFits} returns the formula for the fitted models.
}
\description{
Fitted models for an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}, \code{\linkS4class{edgeSet}}
}
\keyword{modelFits}

