% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{modelFits}
\alias{modelFits}
\alias{modelFits,edgeFit-method}
\title{Fitted models for an \code{\linkS4class{edgeFit}} object.}
\usage{
modelFits(object)

\S4method{modelFits}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}} object}
}
\value{
\code{modelFits} returns the formula for the fitted models.
}
\description{
Fitted models for an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}, \code{\linkS4class{edgeSet}}
}
\keyword{modelFits}

