% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nullMatrix}
\alias{nullMatrix}
\alias{nullMatrix<-}
\title{Matrix representation of null model}
\usage{
nullMatrix(object)

nullMatrix(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{matrix}: null model matrix where columns are covariates and rows are observations}
}
\value{
\code{nullMatrix} returns the value of the null model matrix.
}
\description{
These generic functions access and set the null matrix for
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{fullModel}} and
\code{\link{fullModel}}
}
\keyword{nullMatrix,}
\keyword{nullMatrix<-}

