% Generated by roxygen2 (4.0.1): do not edit by hand
\name{odp}
\alias{odp}
\title{Performs Optimal Discovery Procedure (ODP) on edgeSet object}
\usage{
odp(object, obj.edgeFit, odp.parms = NULL, bs.its = 100, n.mods = 50,
  seed = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object.}

\item{odp.parms}{list- clustering parameters.}

\item{bs.its}{numeric- number of null statistics generated. Default
is 100.}

\item{n.mods}{numeric- number of clusters.}

\item{seed}{numeric- set the seed value.}

\item{verbose}{boolean- print iterations for bootstrap method. Default is TRUE.}

\item{...}{Additional arguments for \code{klClust}, \code{qvalue} and \code{empPvals}.}
}
\value{
\code{odp} returns an \code{\linkS4class{edgeSet}} object with slot qvalueObj as
\code{\link{qvalue}} object.
}
\description{
\code{odp} performs the Optimal Discovery Procedure, which is a new approach for optimally performing
many hypothesis tests in a high-dimensional study. The method has been introduced recently (Storey 2007)
and has been theoretically shown to optimally perform multiple significance tests. When testing a feature,
information from all the features is utilized when testing for significance of a feature. It guarentees to maximize the
number of expected true positive results for each fixed number of expected false positive results which is related to FDR. An
ODP score/statistic is determined for each gene (refer to Storey 2007 for more details) and the null
distribution is generated through a resampling method called bootstrap where residuals from the alternative
model are resampled and added back to the null model to simulate the case where there is no differential
expression.
}
\details{
The full ODP estimator computationally grows quadratically with respect to
the number of genes. This is because the likelihood function of each gene be
calculated across all parameter estimates of the rest of the genes. This becomes
computationally infeasible at a certain point.
Therefore, an alternative method called mODP is used which
has been shown to provide results that are very similar. mODP utilizes a k-means
clustering algorithm where genes are assigned to a cluster based on the Kullback-Leiber distance.
Each gene is assigned an module-average parameter to calculate the ODP score
and it reduces the quadratic time to linear (See Woo, Leek and Storey 2010).
}
\note{
The null and full models are fitted using a linear regression.
To retrieve linear regression results, see \code{\link{edgeFit}}. The p-values
are determined by function \code{\link{empPvals}}.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)

# ODP method
edge.odp <- odp(edgeObj)

# Change the number of clusters
edge.lrt.bs <- lrt(edgeObj, bs.its=10)

# Input an edgeFit object or ODP parameters... not necessary
edge.fit <- edgeFit(edgeObj, stat.type="odp")
edge.parms <- klClust(edgeObj)
edge.odp.ef <- odp(edgeObj, obj.edgeFit=edge.fit, odp.parms=edge.parms)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to simultaneous significance testing. Journal of the Royal Statistical Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for large-scale significance testing, with applications to comparative microarray experiments. Biostatistics, 8: 414-432.

Woo S, Leek JT, Storey JD (2010) A computationally efficient modular optimal discovery procedure. Bioinformatics, 27(4): 509-515.
}
\seealso{
\code{\link{klClust}}, \code{\link{edgeSet}} and \code{\link{edgeFit}}
}
\keyword{odp}

