% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qvalue.obj}
\alias{qvalue.obj}
\alias{qvalue.obj<-}
\title{qvalue object in experiment}
\usage{
qvalue.obj(object, ...)

qvalue.obj(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{S3 \code{object}: \code{\link{qvalue}} object}

\item{\dots}{Additional arguments used in \code{\link{qvalue}}.}
}
\value{
\code{qvalue.obj} returns a \code{\link{qvalue}} object.
}
\description{
These generic functions access and set the \code{qvalue} object in the
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{lrt}}, \code{\link{odp}} and
\code{\linkS4class{edgeSet}}
}
\keyword{qvalue.obj,}
\keyword{qvalue.obj<-}

