% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qvalueObj}
\alias{qvalueObj}
\alias{qvalueObj<-}
\title{qvalue object in experiment}
\usage{
qvalueObj(object)

qvalueObj(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{S3 \code{object}: \code{\link{qvalue}} object}
}
\value{
\code{qvalueObj} returns a \code{\link{qvalue}} object.
}
\description{
These generic functions access and set the \code{qvalue} object in the
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{lrt}}, \code{\link{odp}} and
\code{\linkS4class{edgeSet}}
}
\keyword{qvalueObj,}
\keyword{qvalueObj<-}

