% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resFull}
\alias{resFull}
\title{Residuals of full model fit}
\usage{
resFull(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{resFull} returns a matrix of residuals from full model.
}
\description{
Access the fitted full model residuals in an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{resFull}

