% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{resFull}
\alias{resFull}
\alias{resFull,edgeFit-method}
\title{Residuals of full model fit}
\usage{
resFull(object)

\S4method{resFull}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{resFull} returns a matrix of residuals from full model.
}
\description{
Access the fitted full model residuals in an \code{\linkS4class{edgeFit}}
object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{resFull}

