% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeFit-methods.R, R/edgeSet-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,edgeFit-method}
\alias{show,edgeSet-method}
\title{Efficient printing of edge object}
\usage{
show(object)

\S4method{show}{edgeFit}(object)

\S4method{show}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{\dots}{additional parameters}
}
\description{
Printing method
}
\author{
John Storey, Andrew Bass
}
\keyword{show}

