% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sType}
\alias{sType}
\title{Statistical method used in analysis}
\usage{
sType(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{sType} returns the statistic type- either "odp" or "lrt".
}
\description{
Access the statistic type in an \code{\linkS4class{edgeFit}} object. Can either be the Optimal Discovery Procedure (odp) or the likelihood ratio test (lrt).
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}, \code{\linkS4class{edgeFit}} and
\code{\linkS4class{edgeSet}}
}
\keyword{sType}

