% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeFit-methods.R, R/edgeSet-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,edgeFit-method}
\alias{summary,edgeSet-method}
\title{Summary for edge objects}
\usage{
summary(object, ...)

\S4method{summary}{edgeFit}(object)

\S4method{summary}{edgeSet}(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{\dots}{additional parameters}
}
\description{
Summary of edgeFit and edgeSet objects
}
\author{
John Storey, Andrew Bass
}
\keyword{summary}

