% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/form_models.R
\name{build_study}
\alias{build_study}
\title{Creates an deSet object with appropriate models for users}
\usage{
build_study(data, grp = NULL, adj.var = NULL, bio.var = NULL,
  tme = NULL, ind = NULL, sampling = c("static", "timecourse"),
  basis.df = 2, basis.type = c("ncs", "poly"))
}
\arguments{
\item{data}{\code{matrix}: gene expression data.}

\item{grp}{\code{vector}: groups or biological variable in experiment. Optional.}

\item{adj.var}{\code{matrix}: adjustment variables. Optional.}

\item{bio.var}{\code{matrix}: biological variables. Optional.}

\item{tme}{\code{vector}: covariate of interest in time course study. Optional.}

\item{ind}{\code{factor}: individual factor for repeated observations of the same
individuals. Optional.}

\item{sampling}{\code{string}: type of experiment. Either "static" or
"timecourse". Default is "static".}

\item{basis.df}{\code{numeric}: degree of freedom of the spline fit for time
course study. Default is 2.}

\item{basis.type}{\code{string}: either "ncs" or "ps" basis for time course study.
Default is "ncs".}
}
\value{
\code{build_study} returns an \code{\linkS4class{deSet}} object
with the following slots assigned:
  \describe{
    \item{\code{full.model:}}{full model equation}
   \item{\code{null.model:}}{null model equation}
   \item{\code{full.matrix:}}{full model in matrix form}
   \item{\code{null.matirx:}}{null model in matrix form}
   \item{\code{individual:}}{individuals in experiment (factor)}
   \item{\code{ExpressionSet:}}{inherits ExpressionSet object (assayData,
   phenoData) created in function}
 }
}
\description{
\code{build_study} is a function to create a deSet object without an
ExpressionSet. Alternative and null models are created based on the experiment
type: Either "static" or "timecourse". For more details, refer to the user
manual.
}
\examples{
# create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
age <- kidney$age
sex <- kidney$sex
kidexpr <- kidney$kidexpr

# create deSet object from data
de_obj <- build_study(data = kidexpr, adj.var = sex, tme = age,
sampling = "timecourse", basis.df = 4)
}
\author{
John Storey, Andy Bass
}
\seealso{
\code{\link{deSet}}
}

