% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edge.R
\docType{package}
\name{edge}
\alias{edge}
\alias{edge-package}
\title{Extraction of Differential Gene Expression}
\description{
The edge package implements methods for carrying out differential
expression analyses of genome-wide gene expression studies. Significance
testing using the optimal discovery procedure and generalized likelihood
ratio tests (equivalent to F-tests and t-tests) are implemented for general study
designs. Special functions are available to facilitate the analysis of
common study designs, including time course experiments. Other packages
such as snm, sva, and qvalue are integrated in edge to provide a wide range
of tools for gene expression analysis.
}
\examples{
\dontrun{
browseVignettes("edge")
}
}
\author{
John Storey, Jeffrey Leek, Andrew Bass
}

