% Generated by roxygen2 (4.0.1): do not edit by hand
\name{edgeModel}
\alias{edgeModel}
\title{Creates an edgeSet object and formulates appropriate models for users}
\usage{
edgeModel(data, sampling = c("static", "timecourse"), grp = NULL,
  tme = NULL, ind = NULL, basis.df = 2, basis.type = c("ncs", "poly"),
  bio.var = NULL, adj.var = NULL)
}
\arguments{
\item{data}{matrix- Gene expression data.}

\item{sampling}{string- Type of experiment. Either "static" or "timecourse". Default is "static".}

\item{grp}{vector- Grouping variable in experiment. Optional.}

\item{tme}{vector- Covariate of interest in time course study. Optional.}

\item{ind}{factor- Individual factor. Optional.}

\item{basis.df}{numeric- Degree of freedom of the spline fit for time course study. Default is 2.}

\item{basis.type}{string- Either "ncs" or "ps" basis for time course study. Default is "ncs".}

\item{bio.var}{matrix- Biological variables. Optional.}

\item{adj.var}{matrix- Adjustment Variables. Optional.}
}
\value{
\code{edgeModel} returns an \code{\linkS4class{edgeSet}} object with the following slots assigned:
  \describe{
    \item{\code{full.model:}}{alternative model equation}
   \item{\code{null.model:}}{null model equation}
   \item{\code{full.matrix:}}{alternative model in matrix form}
   \item{\code{null.matirx:}}{null model in matrix form}
   \item{\code{individual:}}{individuals in experiment (factor)}
   \item{\code{ExpressionSet:}}{inherits ExpressionSet object (assayData, phenoData) created in function}
 }
}
\description{
\code{edgeModel} is a function to create an edge object without an ExpressionSet. Alternative and null models are created based on experiment type: Either "static" or "timecourse". For more detail refer to the user manual.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr

#Create edgeSet object from data
edgeObj <- edgeModel(data=kidexpr, adj.var=model.matrix(~sex), tme=age, sampling="timecourse", basis.df=4)
}
\author{
John Storey, Andy Bass
}
\seealso{
\code{\link{edgeSet}}
}

