% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edgeModel.R
\name{edgeModel}
\alias{edgeModel}
\title{Generate alternative and null hypothesis in edgeSet object}
\usage{
edgeModel(data, cov, altMod = NULL, nullMod = NULL, ind = NULL,
  weights = NULL)
}
\arguments{
\item{data}{matrix- Gene expression data.}

\item{weights}{matrix- Matrix of weights for each observation. Optional}

\item{bio.var}{matrix- Biological variables.}

\item{adj.var}{matrix- Adjustment Variables. Optional.}
}
\value{
\code{edgeModel} returns an \code{\linkS4class{edgeSet}} object with the following slots assigned:
  \describe{
    \item{\code{full.model:}}{alternative model equation}
   \item{\code{null.model:}}{null model equation}
   \item{\code{full.matrix:}}{alternative model in matrix form}
   \item{\code{null.matirx:}}{null model in matrix form}
   \item{\code{individual:}}{individuals in experiment (factor)}
   \item{\code{ExpressionSet:}}{inherits ExpressionSet object (assayData, phenoData) created in function}
 }
}
\description{
\code{edgeModel} is a function to create an edge object without an ExpressionSet. Alternative and null models are created based adj.var and bio.var variables. Intercept is included in both alternative and null models by default.
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
kidexpr <- kidney$kidexpr
nullMat <- model.matrix(~sex)
altMat <- model.matrix(~ns(age, df=4))

#Create edgeSet object from data
edgeObj <- edgeModel(data=kidexpr, adj.var=nullMat, bio.var=altMat)
}
\author{
John Storey, Andy Bass
}
\seealso{
\code{\link{edgeSet}}
}

