% Generated by roxygen2 (4.0.1): do not edit by hand
\name{edgeSVA}
\alias{edgeSVA}
\title{Estimate surrogate variables from data based on models in edgeSet object}
\usage{
edgeSVA(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}} object}

\item{...}{Additional arguments for \code{\link{sva}}}
}
\value{
\code{edgeSVA} returns an \code{\linkS4class{edgeSet}} object.
}
\description{
Runs \code{sva} on edgeSet object based on the null and full models in \code{\linkS4class{edgeSet}}
}
\examples{
# Create ExpressionSet object from kidney dataset
library(splines)
data(kidney)
sex <- kidney$sex[kidney$tissue == "c"]
age <- kidney$age[kidney$tissue == "c"]
kidexpr <- log(kidney$kidexpr[, kidney$tissue == "c"] + 10)
expSet <- ExpressionSet(assayData = kidexpr, phenoData = as(data.frame(sex = sex, age = age), "AnnotatedDataFrame"))

# Create Models
nModel <- ~sex
fModel <- ~sex+ ns(age, df=3, intercept=FALSE)

# Create edgeSet
edgeObj <- edgeSet(expSet, full.model=fModel, null.model=nModel)
edgeObj.sva <- edgeSVA(edgeObj)
}
\author{
John Storey, Andrew Bass
}
\references{
Leek JT, Storey JD (2007) Capturing Heterogeneity in Gene Expression Studies by Surrogate Variable Analysis. PLoS Genet 3(9): e161. doi:10.1371/journal.pgen.0030161
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{odp}} and \code{\link{lrt}}
}
\keyword{edgeSVA}

