% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{fitNull}
\alias{fitNull}
\alias{fitNull,edgeFit-method}
\title{Fitted data from the null model}
\usage{
fitNull(object)

\S4method{fitNull}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{fitFull} returns a matrix of fitted values from null model.
}
\description{
Access the fitted data from the null model in an
\code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{fitNull}

