% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{fullMatrix}
\alias{fullMatrix}
\alias{fullMatrix,edgeSet-method}
\alias{fullMatrix<-}
\alias{fullMatrix<-,edgeSet-method}
\title{Matrix representation of full model}
\usage{
fullMatrix(object)

fullMatrix(object) <- value

\S4method{fullMatrix}{edgeSet}(object)

\S4method{fullMatrix}{edgeSet}(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{matrix}: full model matrix where the columns are the
covariates amd rows are observations}
}
\value{
\code{fullMatrix} returns the value of the full model matrix.
}
\description{
These generic functions access and set the full matrix for
\code{\linkS4class{edgeSet}} object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# extract the full model equation as a matrix
mat_full <- fullMatrix(edge_obj)
}
\author{
Andrew Bass, John Storey
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{fullModel}}
}
\keyword{fullMatrix,}
\keyword{fullMatrix<-}

