% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{fullModel}
\alias{fullModel}
\alias{fullModel,edgeSet-method}
\alias{fullModel<-}
\alias{fullModel<-,edgeSet-method}
\title{Full model equation}
\usage{
fullModel(object)

fullModel(object) <- value

\S4method{fullModel}{edgeSet}(object)

\S4method{fullModel}{edgeSet}(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{formula}: The experiment design for the full model.}
}
\value{
\code{fullModel} returns the formula for the full model.
}
\description{
These generic functions access and set the full model for
\code{\linkS4class{edgeSet}} object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# extract out the full model equation
mod_full <- fullModel(edge_obj)

# change the full model in the experiment
fullModel(edge_obj) <- ~sex + ns(age, df = 2)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{fullModel,}
\keyword{fullModel<-}

