% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{fullModel}
\alias{fullModel}
\alias{fullModel,edgeSet-method}
\alias{fullModel<-}
\alias{fullModel<-,edgeSet-method}
\title{Full model equation}
\usage{
fullModel(object)

fullModel(object) <- value

\S4method{fullModel}{edgeSet}(object) <- value

\S4method{fullModel}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{formula}: The experiment design for the full model.}
}
\value{
\code{fullModel} returns the formula for the full model.
}
\description{
These generic functions access and set the full model for
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{fullModel,}
\keyword{fullModel<-}

