% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edge.R
\docType{data}
\name{gibson}
\alias{gibson}
\title{Gene expression dataset from Idaghdour et al. (2008)}
\format{A list called \code{gibson} containing:
\itemize{
  \item batch: Batches in experiment.
  \item location: Location of Moroccan Amazigh groups.
  \item gender: Gender of individuals.
  \item gibexpr: A 500 rows by 46 columns matrix of gene expression values.
}}
\usage{
data(gibson)
}
\description{
The data provide gene expression measurements in peripheral blood leukocyte
samples from three Moroccan Amazigh groups leading distinct ways of life:
desert nomadic (DESERT), mountain agrarian (VILLAGE), and coastal urban
(AGADIR).
}
\note{
These data are a random subset of 500 genes from the total number of genes
in the original data set. To download the full data set, go to
\url{http://genomine.org/edge/}.
}
\examples{
# import
data(gibson)

# create an ExpressionSet
covar <- data.frame(Batch = gibson$batch, Gender = gibson$gender,
Location = gibson$location)
pDat <- as(covar, "AnnotatedDataFrame")
expSet <- ExpressionSet(assayData = gibson$gibexpr, phenoData = pDat)

# create edgeSet for experiment- static experiment
mNull <- ~Gender + Batch
mFull <- ~Gender + Batch + Location

# create edgeSet object
edge_obj <- edgeSet(expSet, full.model = mFull, null.model = mNull)

# Perform ODP/lrt statistic to determine significant genes in study
edge_odp <- odp(edge_obj, bs.its = 10)
edge_lrt <- lrt(edge_obj, nullDistn = "bootstrap", bs.its = 10)

# summarize significance results
summary(edge_odp)
}
\references{
Idaghdour Y, Storey JD, Jadallah S, and Gibson G. (2008) A genome-wide gene
expression signature of lifestyle in peripheral blood of Moroccan Amazighs.
PLoS Genetics, 4: e1000052.
}
\keyword{datasets}

