% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deSet-methods.R
\docType{methods}
\name{lrt}
\alias{lrt}
\alias{lrt,deSet,deFit-method}
\alias{lrt,deSet,missing-method}
\title{Performs F-test (likelihood ratio test using Normal likelihood)}
\usage{
lrt(object, de.fit, nullDistn = c("normal", "bootstrap"), bs.its = 100,
  seed = NULL, verbose = TRUE, ...)

\S4method{lrt}{deSet,missing}(object, de.fit, nullDistn = c("normal",
  "bootstrap"), bs.its = 100, seed = NULL, verbose = TRUE, ...)

\S4method{lrt}{deSet,deFit}(object, de.fit, nullDistn = c("normal",
  "bootstrap"), bs.its = 100, seed = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{S4 object}: \code{\linkS4class{deSet}}}

\item{de.fit}{\code{S4 object}: \code{\linkS4class{deFit}}}

\item{nullDistn}{\code{character}: either "normal" or "bootstrap", If "normal"
then the p-values are calculated using an F-test. If "bootstrap" then
p-values will be determined from the function \code{\link{empPvals}}.
Default is "normal".}

\item{bs.its}{\code{integer}: number of null statistics generated (only applicable
for "bootstrap" method. Default is 100.)}

\item{seed}{\code{integer}: set the seed value. Default is NULL.}

\item{verbose}{\code{boolean}: print iterations for bootstrap method.
Default is TRUE.}

\item{...}{Additional arguments for \code{\link{apply_qvalue}} and
\code{\link{empPvals}}
function.}
}
\value{
\code{lrt} returns a \code{\linkS4class{deSet}} object
}
\description{
\code{lrt} performs a generalized likelihood ratio test using the full and
null models to determine p-values (either from a F-distribution or from
simulated null test statistics). If the null distribution is calculated
using "bootstrap" then residuals from the alternative model are resampled
and added to the null model residuals to simulate a distribution where
there is no differential expression. Otherwise, the default input is
"normal" and the assumption is that the dataset follows an F-distribution.
}
\note{
Fits the full and null models to each gene using the function
\code{\link{fit_models}} and then performs an ANOVA F-test. If \code{nullDistn}
"bootstrap" is chosen then empirical p-values will
be determined from the \code{\link{qvalue}} package (see
\code{\link{empPvals}}).
}
\examples{
# import data
library(splines)
data(kidney)
age <- kidney$age
sex <- kidney$sex
kidexpr <- kidney$kidexpr
cov <- data.frame(sex = sex, age = age)

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create deSet object from data
de_obj <- build_models(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# lrt method
de_lrt <- lrt(de_obj, nullDistn = "normal")

# to generate p-values from bootstrap
de_lrt <- lrt(de_obj, nullDistn = "bootstrap", bs.its = 30)

# input an deFit object but not necessary
de_fit <- fit_models(de_obj, stat.type = "lrt")
de_lrt <- lrt(de_obj, de.fit = de_fit)
summary(de_lrt)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD, Xiao W, Leek JT, Tompkins RG, and Davis RW. (2005) Significance
analysis of time course microarray experiments. Proceedings of the National
Academy of Sciences, 102: 12837-12842.
}
\seealso{
\code{\link{deSet}}, \code{\link{odp}}, \code{\link{build_models}}
}

