% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{lrt}
\alias{lrt}
\alias{lrt,edgeSet,edgeFit-method}
\alias{lrt,edgeSet,missing-method}
\title{Performs F-test (likelihood ratio test using Normal likelihood)}
\usage{
lrt(object, obj.edgeFit, nullDistn = c("normal", "bootstrap"), bs.its = 100,
  seed = NULL, verbose = TRUE, ...)

\S4method{lrt}{edgeSet,missing}(object, obj.edgeFit, nullDistn = c("normal",
  "bootstrap"), bs.its = 100, seed = NULL, verbose = TRUE, ...)

\S4method{lrt}{edgeSet,edgeFit}(object, obj.edgeFit, nullDistn = c("normal",
  "bootstrap"), bs.its = 100, seed = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object- S4 class}

\item{nullDistn}{character- either "normal" or "bootstrap", If "normal",
use p-values from ANOVA else if "bootstrap" p-values will
be determined from function \code{\link{empPvals}}.
Default is "normal".}

\item{bs.its}{numeric- number of null statistics generated (only applicable
for "bootstrap" method. Default is 100.)}

\item{seed}{numeric- set the seed value.}

\item{verbose}{boolean- print iterations for bootstrap method. Default is
TRUE.}

\item{...}{Additional arguments for \code{qvalue} and \code{empPvals}
function.}
}
\value{
\code{lrt} returns a \code{\linkS4class{edgeSet}} object with slot
qvalueObj as a \code{\link{qvalue}} object.
}
\description{
\code{lrt} performs a generalized likelihood ratio test using the full and
null models to determine p-values (either from a F-distribution or from
simulated null test statistics). If the null distribution is calculated
using "bootstrap" then residuals from the alternative model are resampled
and added to the null model to simulate a distribution where there is no
differential expression. Otherwise, the default input is "normal" and the
assumption is that the dataset follows an F-distribution. Note that
if "normal" is chosen then the sum of squares should be statistically
independent and genes should be normally distributed with the same
variance. The \code{qvalue} function is used to retrieve the \code{qvalue}
object. See \code{\link{qvalue}} for more information.
}
\note{
Fits a linear regression to each gene from function
\code{\link{edgeFit}} and then performs an ANOVA F-test on the null and full
model fits. If nullDistn "bootstrap" is chosen then empirical p-values will
be determined from the \code{\link{qvalue}} package (see
\code{\link{empPvals}}). To retrieve linear regression results, see
\code{\link{edgeFit}}.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# lrt method
edge_lrt <- lrt(edge_obj, nullDistn = "normal")

# to generate p-values from bootstrap
edge_lrt <- lrt(edge_obj, nullDistn = "bootstrap", bs.its=30)

# input an edgeFit object but not necessary
edge_fit <- edgeFit(edge_obj, stat.type = "lrt")
edge_lrt <- lrt(edge_obj, obj.edgeFit = edge_fit)
}
\author{
John Storey, Andrew Bass
}
\references{
Storey JD, Xiao W, Leek JT, Tompkins RG, and Davis RW. (2005) Significance
analysis of time course microarray experiments. Proceedings of the National
Academy of Sciences, 102: 12837-12842.
}
\seealso{
\code{\link{edgeSet}}, \code{\link{odp}}, \code{\link{edgeFit}}
}
\keyword{lrt}

