% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{nullMatrix}
\alias{nullMatrix}
\alias{nullMatrix,edgeSet-method}
\alias{nullMatrix<-}
\alias{nullMatrix<-,edgeSet-method}
\title{Matrix representation of null model}
\usage{
nullMatrix(object)

nullMatrix(object) <- value

\S4method{nullMatrix}{edgeSet}(object) <- value

\S4method{nullMatrix}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{matrix}: null model matrix where columns are covariates
and rows are observations}
}
\value{
\code{nullMatrix} returns the value of the null model matrix.
}
\description{
These generic functions access and set the null matrix for
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{fullModel}} and
\code{\link{fullModel}}
}
\keyword{nullMatrix,}
\keyword{nullMatrix<-}

