% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{nullMatrix}
\alias{nullMatrix}
\alias{nullMatrix,edgeSet-method}
\alias{nullMatrix<-}
\alias{nullMatrix<-,edgeSet-method}
\title{Matrix representation of null model}
\usage{
nullMatrix(object)

nullMatrix(object) <- value

\S4method{nullMatrix}{edgeSet}(object)

\S4method{nullMatrix}{edgeSet}(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{matrix}: null model matrix where columns are covariates
and rows are observations}
}
\value{
\code{nullMatrix} returns the value of the null model matrix.
}
\description{
These generic functions access and set the null matrix for
\code{\linkS4class{edgeSet}} object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# extract the null model as a matrix
mat_null <- nullMatrix(edge_obj)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}, \code{\link{fullModel}} and
\code{\link{fullModel}}
}
\keyword{nullMatrix,}
\keyword{nullMatrix<-}

