% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{nullModel}
\alias{nullModel}
\alias{nullModel,edgeSet-method}
\alias{nullModel<-}
\alias{nullModel<-,edgeSet-method}
\title{Null model equation from edgeSet object}
\usage{
nullModel(object)

nullModel(object) <- value

\S4method{nullModel}{edgeSet}(object)

\S4method{nullModel}{edgeSet}(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{formula}: The experiment design for the null model.}
}
\value{
\code{nullModel} returns the formula for the null model.
}
\description{
These generic functions access and set the null model for
\code{\linkS4class{edgeSet}} object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# extract the null model equation
mod_null <- nullModel(edge_obj)

# change null model in experiment but must update full model
nullModel(edge_obj) <- ~1
fullModel(edge_obj) <- ~1 + ns(age, df=4)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{nullModel,}
\keyword{nullModel<-}

