% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{nullModel}
\alias{nullModel}
\alias{nullModel,edgeSet-method}
\alias{nullModel<-}
\alias{nullModel<-,edgeSet-method}
\title{Null model equation from edgeSet object}
\usage{
nullModel(object)

nullModel(object) <- value

\S4method{nullModel}{edgeSet}(object) <- value

\S4method{nullModel}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{\code{formula}: The experiment design for the null model.}
}
\value{
\code{nullModel} returns the formula for the null model.
}
\description{
These generic functions access and set the null model for
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\linkS4class{edgeSet}}
}
\keyword{nullModel,}
\keyword{nullModel<-}

