% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeSet-methods.R
\docType{methods}
\name{odp}
\alias{odp}
\alias{odp,edgeSet,edgeFit-method}
\alias{odp,edgeSet,missing-method}
\title{Performs optimal discovery procedure (ODP) on an edgeSet object}
\usage{
odp(object, obj.edgeFit, odp.parms = NULL, bs.its = 100, n.mods = 50,
  seed = NULL, verbose = TRUE, ...)

\S4method{odp}{edgeSet,missing}(object, obj.edgeFit, odp.parms = NULL,
  bs.its = 100, n.mods = 50, seed = NULL, verbose = TRUE, ...)

\S4method{odp}{edgeSet,edgeFit}(object, obj.edgeFit, odp.parms = NULL,
  bs.its = 100, n.mods = 50, seed = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{obj.edgeFit}{edgeFit object.}

\item{odp.parms}{list- clustering parameters.}

\item{bs.its}{numeric- number of null statistics generated. Default
is 100.}

\item{n.mods}{numeric- number of clusters.}

\item{seed}{numeric- set the seed value.}

\item{verbose}{boolean- print iterations for bootstrap method. Default is
TRUE.}

\item{...}{Additional arguments for \link{\code{klClust}}, \code{qvalue} and
\code{empPvals} (latter two from \code{qvalue} package).}
}
\value{
\code{odp} returns an \code{\linkS4class{edgeSet}} object with slot
qvalueObj as a \code{\link{qvalue}} object.
}
\description{
\code{odp} performs the optimal discovery procedure, which is a new
approach for optimally performing many hypothesis tests in a
high-dimensional study. When testing whether a feature is significant, the
optimal discovery procedure uses information across all features when
testing for significance. \cr
}
\details{
It guarentees to maximize the number of expected true positive results for
each fixed number of expected false positive results- which is related to
false discovery rates.

The full ODP estimator computationally grows quadratically with
respect to the number of genes. This becomes computationally infeasible at
a certain point. Therefore, an alternative method called mODP is used which
has been shown to provide results that are very similar. mODP utilizes a
k-means clustering algorithm where genes are assigned to a cluster based on
the Kullback-Leiber distance. Each gene is assigned an module-average
parameter to calculate the ODP score and it reduces the quadratic time to
linear (See Woo, Leek and Storey 2010).
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# odp method
edge_odp <- odp(edge_obj, bs.its = 30)

# input an edgeFit object or ODP parameters... not necessary
edge_fit <- edgeFit(edge_obj, stat.type = "odp")
edge_clust <- klClust(edge_obj, n.mods = 10)
edge_odp <- odp(edge_obj, obj.edgeFit = edge_fit, odp.parms = edge_clust,
bs.its = 30)
}
\author{
John Storey, Jeffrey Leek, Andrew Bass
}
\references{
Storey JD. (2007) The optimal discovery procedure: A new approach to
simultaneous significance testing. Journal of the Royal Statistical
Society, Series B, 69: 347-368.

Storey JD, Dai JY, and Leek JT. (2007) The optimal discovery procedure for
large-scale significance testing, with applications to comparative
microarray experiments. Biostatistics, 8: 414-432.

Woo S, Leek JT, Storey JD (2010) A computationally efficient modular
optimal discovery procedure. Bioinformatics, 27(4): 509-515.
}
\seealso{
\code{\link{klClust}}, \code{\link{edgeSet}} and
\code{\link{edgeFit}}
}
\keyword{odp}

