% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R, R/setMethods.R
\docType{methods}
\name{qvalueObj}
\alias{qvalueObj}
\alias{qvalueObj,edgeSet-method}
\alias{qvalueObj<-}
\alias{qvalueObj<-,edgeSet-method}
\title{Access/set qvalue slot}
\usage{
qvalueObj(object)

qvalueObj(object) <- value

\S4method{qvalueObj}{edgeSet}(object) <- value

\S4method{qvalueObj}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{value}{S3 \code{object}: \code{\link{qvalue}} object}
}
\value{
\code{qvalueObj} returns a \code{\link{qvalue}} object.
}
\description{
These generic functions access and set the \code{qvalue} object in the
\code{\linkS4class{edgeSet}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{lrt}}, \code{\link{odp}} and
\code{\linkS4class{edgeSet}}
}
\keyword{qvalueObj,}
\keyword{qvalueObj<-}

