% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{resFull}
\alias{resFull}
\alias{resFull,edgeFit-method}
\title{Residuals of full model fit}
\usage{
resFull(object)

\S4method{resFull}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{resFull} returns a matrix of residuals from full model.
}
\description{
Access the fitted full model residuals in an \code{\linkS4class{edgeFit}}
object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# run edgeFit to get model fits
edge_fit <- edgeFit(edge_obj)

# extract out the full residuals from the model fit
res_full <- resFull(edge_fit)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{resFull}

