% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resNull}
\alias{resNull}
\title{Residuals of null model fit}
\usage{
resNull(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{resNull} returns a matrix of residuals from null model.
}
\description{
Access the fitted null model residuals in an \code{\linkS4class{edgeFit}} object.
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{resNull}

