% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{resNull}
\alias{resNull}
\alias{resNull,edgeFit-method}
\title{Residuals of null model fit}
\usage{
resNull(object)

\S4method{resNull}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{resNull} returns a matrix of residuals from null model.
}
\description{
Access the fitted null model residuals in an \code{\linkS4class{edgeFit}}
object.
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# run edgeFit to get model fits
edge_fit <- edgeFit(edge_obj)

# extract out the null residuals from the model fits
res_null <- resNull(edge_fit)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}
}
\keyword{resNull}

