% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeFit-methods.R, R/edgeSet-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,edgeFit-method}
\alias{show,edgeSet-method}
\title{Efficient printing of edge object}
\usage{
show(object)

\S4method{show}{edgeFit}(object)

\S4method{show}{edgeSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{\dots}{additional parameters}
}
\value{
show returns an invisible NULL.
}
\description{
Printing method
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# show method
edge_obj
}
\author{
John Storey, Andrew Bass
}
\keyword{show}

