% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{sType}
\alias{sType}
\alias{sType,edgeFit-method}
\title{Statistical method used in analysis}
\usage{
sType(object)

\S4method{sType}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{sType} returns the statistic type- either "odp" or "lrt".
}
\description{
Access the statistic type in an \code{\linkS4class{edgeFit}} object. Can
either be the optimal discovery procedure (odp) or the likelihood ratio
test (lrt).
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}, \code{\linkS4class{edgeFit}} and
\code{\linkS4class{edgeSet}}
}
\keyword{sType}

