% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMethods.R
\docType{methods}
\name{sType}
\alias{sType}
\alias{sType,edgeFit-method}
\title{Statistical method used in analysis}
\usage{
sType(object)

\S4method{sType}{edgeFit}(object)
}
\arguments{
\item{object}{\code{\linkS4class{edgeFit}}}
}
\value{
\code{sType} returns the statistic type- either "odp" or "lrt".
}
\description{
Access the statistic type in an \code{\linkS4class{edgeFit}} object. Can
either be the optimal discovery procedure (odp) or the likelihood ratio
test (lrt).
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null.model <- ~sex
full.model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, nullMod = null.model,
altMod = full.model)

# run edgeFit to get model fits
edge_fit <- edgeFit(edge_obj)

# extract the statistic type of model fits
stat_type <- sType(edge_fit)
}
\author{
John Storey, Andrew Bass
}
\seealso{
\code{\link{edgeFit}}, \code{\linkS4class{edgeFit}} and
\code{\linkS4class{edgeSet}}
}
\keyword{sType}

