% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edgeFit-methods.R, R/edgeSet-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,edgeFit-method}
\alias{summary,edgeSet-method}
\title{Summary for edge objects}
\usage{
summary(object, ...)

\S4method{summary}{edgeFit}(object)

\S4method{summary}{edgeSet}(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{\dots}{additional parameters}
}
\value{
Summary of edgeSet object
}
\description{
Summary of edgeFit and edgeSet objects
}
\examples{
# import data
library(splines)
data(kidney)
sex <- kidney$sex
age <- kidney$age
cov <- data.frame(sex = sex, age = age)
kidexpr <- kidney$kidexpr

# create models
null_model <- ~sex
full_model <- ~sex + ns(age, df = 4)

# create edgeSet object from data
edge_obj <- edgeModel(data = kidexpr, cov = cov, null.model = null_model,
full.model = full_model)

# get summary
summary(edge_obj)

# run odp and summarize
edge_odp <- odp(edge_obj, bs.its= 20)
summary(edge_odp)
}
\author{
John Storey, Andrew Bass
}
\keyword{summary}

