% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary}
\alias{summary}
\title{Summary of edge objects}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{edgeSet}}}

\item{\dots}{additional parameters}
}
\description{
Summary of edgeFit and edgeSet objects
}
\author{
John Storey, Andrew Bass
}
\keyword{summary}

